# docs/api-reference/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/apps/{id}": {
      "get": {
        "operationId": "AppsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Get an app with its latest snapshot and competitors",
        "tags": [
          "apps"
        ]
      },
      "delete": {
        "operationId": "AppsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete an app",
        "tags": [
          "apps"
        ]
      }
    },
...
    "/apps/{id}/keywords": {
      "get": {
        "operationId": "KeywordsController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "opportunity",
                "traffic",
                "difficulty",
                "position",
                "volatility"
              ]
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "schema": {
              "pattern": "^[a-z]{2}$",
              "example": "pl",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "List tracked keywords for an app",
        "tags": [
          "keywords"
        ]
      },
      "post": {
        "operationId": "KeywordsController_add",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddKeywordsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "Add manual keywords",
        "tags": [
          "keywords"
        ]
      }
    },
    "/apps/{id}/keyword-countries": {
      "get": {
        "operationId": "KeywordsController_countries",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "List the markets an app tracks keywords in",
        "tags": [
          "keywords"
        ]
      }
    },
    "/apps/{id}/keywords/compare": {
      "get": {
        "operationId": "KeywordsController_compare",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "onlyGaps",
            "required": false,
            "in": "query",
            "schema": {
              "default": false,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Compare keyword positions against competitors",
        "tags": [
          "keywords"
        ]
      }
    },
    "/apps/{id}/keywords/suggestions": {
      "get": {
        "operationId": "KeywordsController_suggestions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "strategy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "metadata",
              "type": "string",
              "enum": [
                "metadata",
                "search",
                "similar",
                "developer",
                "competitors",
                "seasonal",
                "reviews"
              ]
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "schema": {
              "pattern": "^[a-z]{2}$",
              "example": "pl",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 30,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "Suggest keywords via metadata, search or similar",
        "tags": [
          "keywords"
        ]
      }
    },
    "/apps/{id}/keywords/spider": {
      "get": {
        "operationId": "KeywordsController_spiderStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "term",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 2,
              "maxLength": 40,
              "example": "habit tracker",
              "type": "string"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "schema": {
              "pattern": "^[a-z]{2}$",
              "example": "pl",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Aggregate suggest spider probes for a term",
        "tags": [
          "keywords"
        ]
      },
      "post": {
        "operationId": "KeywordsController_startSpider",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpiderStartDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Enqueue suggest spider probes for a term",
        "tags": [
          "keywords"
        ]
      }
    },
    "/apps/{id}/keywords/{keywordId}": {
      "patch": {
        "operationId": "KeywordsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keywordId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateKeywordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Update a tracked keyword active flag or relevance",
        "tags": [
          "keywords"
        ]
      },
      "delete": {
        "operationId": "KeywordsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keywordId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Stop tracking a keyword",
        "tags": [
          "keywords"
        ]
      }
    },
    "/apps/{id}/keyword-field": {
      "put": {
        "operationId": "KeywordsController_setKeywordField",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeywordFieldDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Set the manual iOS keyword field",
        "tags": [
          "keywords"
        ]
      }
    },
...
    "/apps/{id}/reviews": {
      "get": {
        "operationId": "ReviewsController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "score",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 5,
              "type": "number"
            }
          },
          {
            "name": "version",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Stored reviews for an app, filtered by star and version",
        "tags": [
          "reviews"
        ]
      }
    },
    "/apps/{id}/reviews/histogram": {
      "get": {
        "operationId": "ReviewsController_histogram",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Ratings distribution from the latest snapshot (Google Play only)",
        "tags": [
          "reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "asobeast API",
    "description": "Self hosted ASO toolkit API.\n\nEvery response carries RateLimit-Limit, RateLimit-Remaining and RateLimit-Reset for the tightest limit the request touched. Limits are per workspace, so minting more tokens does not buy more capacity, and they are sized per plan across three classes: read for stored data, write for configuration changes and store for the endpoints that reach the App Store or Google Play.\n\nA 429 carries Retry-After and a rateLimit object naming the window that closed. A 403 with a quota object means a plan capacity limit such as apps or tracked keyword markets, which is a different problem from a rate limit. A 402 with an entitlement object means the workspace needs a plan.\n\nPositions are 1-based and null means checked but not found within the captured depth, never zero. Daily granularity uses UTC dates.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/backend",
      "description": "Web origin proxy"
    },
    {
      "url": "http://localhost:4000",
      "description": "Direct API"
    }
  ],
  "components": {
    "securitySchemes": {
      "personalApiToken": {
        "scheme": "bearer",
        "bearerFormat": "asob",
        "type": "http"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "SpiderStartDto": {
        "type": "object",
        "properties": {
          "term": {
            "type": "string",
            "minLength": 2,
            "maxLength": 40,
            "example": "habit tracker"
          },
          "country": {
            "type": "string",
            "example": "pl",
            "pattern": "^[a-z]{2}$"
          }
        },
        "required": [
          "term"
        ]
      },
      "AddKeywordsDto": {
        "type": "object",
        "properties": {
          "keywords": {
            "minItems": 1,
            "example": [
              "habit tracker",
              "streak counter"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "country": {
            "type": "string",
            "example": "pl",
            "pattern": "^[a-z]{2}$"
          }
        },
        "required": [
          "keywords"
        ]
      },
      "UpdateKeywordDto": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "example": false
          },
          "relevance": {
            "type": "number",
            "nullable": true,
            "minimum": 1,
            "maximum": 100,
            "example": 80
          }
        }
      },
      "KeywordFieldDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "habit,tracker,streak,daily goals"
          }
        },
        "required": [
          "text"
        ]
      },
...
    }
  },
  "security": [
    {
      "personalApiToken": []
    }
  ]
}